Make ExchangeSampler only propose different configurations
