+++ netket/experimental/sampler/rules/fermion_2nd.py
@@ -1,7 +1,3 @@
-from functools import partial
-
-import jax
-import jax.numpy as jnp
 from typing import Optional
 import numpy as np
 
@@ -9,7 +5,6 @@ from netket.sampler.rules import ExchangeRule
 from netket.graph import AbstractGraph
 from netket.graph import disjoint_union
 from netket.experimental.hilbert import SpinOrbitalFermions
-from netket.jax.sharding import sharding_decorator
 
 
 class ParticleExchangeRule(ExchangeRule):
@@ -72,52 +67,5 @@ class ParticleExchangeRule(ExchangeRule):
                 )
         super().__init__(clusters=clusters, graph=graph, d_max=d_max)
 
-    def transition(rule, sampler, machine, parameters, state, key, σ):
-        n_chains = σ.shape[0]
-
-        # compute a mask for the clusters that can be hopped
-        hoppable_clusters = _compute_hoppable_clusters_mask(rule.clusters, σ)
-
-        keys = jnp.asarray(jax.random.split(key, n_chains))
-
-        # we use shard_map to avoid the all-gather coming from the batched jnp.take / indexing
-        @partial(sharding_decorator, sharded_args_tree=(True, True, True))
-        @jax.vmap
-        def _update_samples(key, σ, hoppable_clusters):
-            # pick a random cluster, taking into account the mask
-            n_conn = hoppable_clusters.sum(axis=-1)
-            cluster = jax.random.choice(
-                key,
-                a=jnp.arange(rule.clusters.shape[0]),
-                p=hoppable_clusters,
-                replace=True,
-            )
-
-            # sites to be exchanged
-            si = rule.clusters[cluster, 0]
-            sj = rule.clusters[cluster, 1]
-
-            σp = σ.at[si].set(σ[sj])
-            σp = σp.at[sj].set(σ[si])
-
-            # compute the number of connected sites
-            hoppable_clusters_proposed = _compute_hoppable_clusters_mask(
-                rule.clusters, σp
-            )
-            n_conn_proposed = hoppable_clusters_proposed.sum(axis=-1)
-            log_prob_corr = jnp.log(n_conn) - jnp.log(n_conn_proposed)
-            return σp, log_prob_corr
-
-        return _update_samples(keys, σ, hoppable_clusters)
-
     def __repr__(self):
         return f"ParticleExchangeRule(# of clusters: {len(self.clusters)})"
-
-
-@jax.jit
-def _compute_hoppable_clusters_mask(clusters, σ):
-    # mask the clusters to include only feasible moves (occ -> unocc, or the inverse)
-    hoppable_clusters_mask = ~jnp.isclose(
-        σ[..., clusters[:, 0]], σ[..., clusters[:, 1]]
-    )
-    return hoppable_clusters_mask
