+++ netket/sampler/rules/exchange.py
@@ -41,8 +41,9 @@ class ExchangeRule(MetropolisRule):
 
     1. A pair of indices :math:`i,j = 1\dots N`, and such
        that :math:`\mathrm{dist}(i,j) \leq d_{\mathrm{max}}`,
-       is chosen with uniform probability.
-    2. The sites are exchanged, i.e. :math:`s^\prime_i = s_j` and :math:`s^\prime_j = s_i`.
+       is chosen with uniform probability, excluding all the sites where
+       :math:`s_i = s_j`;
+    2. The sites are exchanged, i.e. :math:`s^\prime_i = s_j` and :math:`s^\prime_j = s_i`;
 
     Notice that this sampling method generates random permutations of the quantum
     numbers, thus global quantities such as the sum of the local quantum numbers
@@ -97,23 +98,40 @@ class ExchangeRule(MetropolisRule):
     def transition(rule, sampler, machine, parameters, state, key, σ):
         n_chains = σ.shape[0]
 
-        # pick a random cluster
-        cluster_id = jax.random.randint(
-            key, shape=(n_chains,), minval=0, maxval=rule.clusters.shape[0]
-        )
+        # compute a mask for the clusters that can be hopped
+        hoppable_clusters = _compute_different_clusters_mask(rule.clusters, σ)
 
-        # we use shard_map to avoid the all-gather emitted by the batched jnp.take / indexing
-        @partial(sharding_decorator, sharded_args_tree=(True, True))
+        keys = jnp.asarray(jax.random.split(key, n_chains))
+
+        # we use shard_map to avoid the all-gather coming from the batched jnp.take / indexing
+        @partial(sharding_decorator, sharded_args_tree=(True, True, True))
         @jax.vmap
-        def update_fun(σ, cluster):
-            # sites to be exchanged,
+        def _update_samples(key, σ, hoppable_clusters):
+            # pick a random cluster, taking into account the mask
+            n_conn = hoppable_clusters.sum(axis=-1)
+            cluster = jax.random.choice(
+                key,
+                a=jnp.arange(rule.clusters.shape[0]),
+                p=hoppable_clusters,
+                replace=True,
+            )
+
+            # sites to be exchanged
             si = rule.clusters[cluster, 0]
             sj = rule.clusters[cluster, 1]
 
             σp = σ.at[si].set(σ[sj])
-            return σp.at[sj].set(σ[si])
+            σp = σp.at[sj].set(σ[si])
 
-        return (update_fun(σ, cluster_id), None)
+            # compute the number of connected sites
+            hoppable_clusters_proposed = _compute_different_clusters_mask(
+                rule.clusters, σp
+            )
+            n_conn_proposed = hoppable_clusters_proposed.sum(axis=-1)
+            log_prob_corr = jnp.log(n_conn) - jnp.log(n_conn_proposed)
+            return σp, log_prob_corr
+
+        return _update_samples(keys, σ, hoppable_clusters)
 
     def __repr__(self):
         return f"ExchangeRule(# of clusters: {len(self.clusters)})"
@@ -139,3 +157,12 @@ def compute_clusters(graph: AbstractGraph, d_max: int):
         res_clusters[i] = np.asarray(cluster)
 
     return res_clusters
+
+
+@jax.jit
+def _compute_different_clusters_mask(clusters, σ):
+    # mask the clusters to include only feasible moves (occ -> unocc, or the inverse)
+    hoppable_clusters_mask = ~jnp.isclose(
+        σ[..., clusters[:, 0]], σ[..., clusters[:, 1]]
+    )
+    return hoppable_clusters_mask
